demo-std refactoring
